+++ packages/engine/Specs/Core/PeliasGeocoderServiceSpec.js
@@ -1,18 +1,23 @@
 import {
   Cartesian3,
+  GeocoderService,
   GeocodeType,
   PeliasGeocoderService,
   Resource,
 } from "../../index.js";
 
 describe("Core/PeliasGeocoderService", function () {
+  it("conforms to GeocoderService interface", function () {
+    expect(PeliasGeocoderService).toConformToInterface(GeocoderService);
+  });
+
   it("constructor throws without url", function () {
     expect(function () {
       return new PeliasGeocoderService(undefined);
     }).toThrowDeveloperError();
   });
 
-  it("returns geocoder results", function () {
+  it("returns geocoder results", async function () {
     const service = new PeliasGeocoderService("http://test.invalid/v1/");
 
     const query = "some query";
@@ -34,14 +39,13 @@ describe("Core/PeliasGeocoderService", function () {
       Promise.resolve(data)
     );
 
-    return service.geocode(query).then(function (results) {
-      expect(results.length).toEqual(1);
-      expect(results[0].displayName).toEqual(data.features[0].properties.label);
-      expect(results[0].destination).toBeInstanceOf(Cartesian3);
-    });
+    const results = await service.geocode(query);
+    expect(results.length).toEqual(1);
+    expect(results[0].displayName).toEqual(data.features[0].properties.label);
+    expect(results[0].destination).toBeInstanceOf(Cartesian3);
   });
 
-  it("returns no geocoder results if Pelias has no results", function () {
+  it("returns no geocoder results if Pelias has no results", async function () {
     const service = new PeliasGeocoderService("http://test.invalid/v1/");
 
     const query = "some query";
@@ -50,12 +54,11 @@ describe("Core/PeliasGeocoderService", function () {
       Promise.resolve(data)
     );
 
-    return service.geocode(query).then(function (results) {
-      expect(results.length).toEqual(0);
-    });
+    const results = await service.geocode(query);
+    expect(results.length).toEqual(0);
   });
 
-  it("calls search endpoint if specified", function () {
+  it("calls search endpoint if specified", async function () {
     const service = new PeliasGeocoderService("http://test.invalid/v1/");
 
     const query = "some query";
@@ -68,7 +71,7 @@ describe("Core/PeliasGeocoderService", function () {
       "getDerivedResource"
     ).and.callThrough();
 
-    service.geocode(query, GeocodeType.SEARCH);
+    await service.geocode(query, GeocodeType.SEARCH);
     expect(getDerivedResource).toHaveBeenCalledWith({
       url: "search",
       queryParameters: {
@@ -77,7 +80,7 @@ describe("Core/PeliasGeocoderService", function () {
     });
   });
 
-  it("calls autocomplete endpoint if specified", function () {
+  it("calls autocomplete endpoint if specified", async function () {
     const service = new PeliasGeocoderService("http://test.invalid/v1/");
 
     const query = "some query";
@@ -90,7 +93,7 @@ describe("Core/PeliasGeocoderService", function () {
       "getDerivedResource"
     ).and.callThrough();
 
-    service.geocode(query, GeocodeType.AUTOCOMPLETE);
+    await service.geocode(query, GeocodeType.AUTOCOMPLETE);
     expect(getDerivedResource).toHaveBeenCalledWith({
       url: "autocomplete",
       queryParameters: {