+++ CHANGES.md
@@ -10,6 +10,7 @@
 - Added `CesiumWidget.creditDisplay` to access the onscreen and lightbox credits.
 - Added `CreditDisplay.addStaticCredit` and `CreditDisplay.removeStaticCredit` such that `Credit.showOnScreen` value is taken into account. [#6215](https://github.com/CesiumGS/cesium/issues/6215)
 - Added `options.gltfCallback` to `Model.loadGltfAsync` to allow apps to access the loaded glTF JSON. [#11240](https://github.com/CesiumGS/cesium/pull/11240)
+- Added `GeocoderService.credit` and and `attributions` property to `GeocoderService.Result` to allow for geocoder services to attribute results.
 
 ##### Fixes :wrench:
 
@@ -34,6 +35,7 @@
 ##### Additions :tada:
 
 - Added `Viewer.creditDisplay` to access the onscreen and lightbox credits.
+- The `Geocoder` widget will now display attributions onscreen or in the lightbox for geocoder results if present, otherwise a default credit from a geocoder service if one is provided.
 
 ##### Fixes :wrench: